feat: 작성자 분리에 따른 CRUD 기존코드 변경 완료
해당 사항들에 맞는 에러 메세지 처리 완료

+++ src/main/java/com/example/springschedulerapi/service/ScheduleServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.springschedulerapi.service;
 
 import com.example.springschedulerapi.dao.ScheduleDao;
+import com.example.springschedulerapi.exception.PasswordMissMatchException;
+import com.example.springschedulerapi.exception.ScheduleNotFoundException;
 import com.example.springschedulerapi.model.dto.request.ScheduleRequestDTO;
 import com.example.springschedulerapi.model.dto.response.ScheduleResponseDTO;
 import lombok.RequiredArgsConstructor;
@@ -8,63 +10,81 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ScheduleServiceImpl implements ScheduleService {
 
     private final ScheduleDao scheduleDao;
+    private final AuthorService authorService;
 
     @Override
-    public ScheduleResponseDTO saveSchedule(ScheduleRequestDTO dto) {
+    public ScheduleResponseDTO createSchedule(ScheduleRequestDTO dto) {
+        //사용자 검증
+        authorService.validateAuthor(dto);
         Long id = scheduleDao.insertSchedule(dto);
-        return scheduleDao.getScheduleById(id);
+        return scheduleDao.getScheduleById(id).get();
     }
 
     @Override
-    public List<ScheduleResponseDTO> searchSchedules(LocalDate start, LocalDate end, String author) {
+    public List<ScheduleResponseDTO> searchSchedules(LocalDate start, LocalDate end, Long authorId) {
 
-        //날짜에 시간추가
-        LocalDateTime startTime = start.atStartOfDay();
-        LocalDateTime endTime = end.plusDays(1).atStartOfDay();
+        boolean dateCondition = (start != null && end != null);
+        boolean authorCondition = authorId != null;
 
-        List<ScheduleResponseDTO> schedules;
-
-        if(author == null || author.isEmpty())
-            return scheduleDao.getSchedulesByDay(startTime, endTime);
+        if(!dateCondition && !authorCondition)
+            return scheduleDao.getAllSchedule();
+        else if(!dateCondition && authorCondition)
+            return scheduleDao.getScheduleByAuthor(authorId);
+        else if(dateCondition && !authorCondition)
+            return scheduleDao.getSchedulesByDay(start, end.plusDays(1));
         else
-            return scheduleDao.getSchedulesByDayAndAuthor(startTime, endTime, author);
+            return scheduleDao.getSchedulesByDayAndAuthor(start, end.plusDays(1), authorId);
     }
 
     @Override
     public ScheduleResponseDTO searchSchedule(Long id) {
-        return scheduleDao.getScheduleById(id);
-    }
-
-    @Override
-    public boolean checkPassword(Long id, ScheduleRequestDTO dto) {
-        if (dto.getPassword().equals(scheduleDao.getPasswordById(id))){
-            return true;
-        }
-        return false;
+        Optional<ScheduleResponseDTO> optional = scheduleDao.getScheduleById(id);
+        if(optional.isPresent())
+            return optional.get();
+        else
+            throw new ScheduleNotFoundException(id + "은 없는 아이디 입니다.");
     }
 
     @Transactional
     @Override
     public ScheduleResponseDTO updateSchedule(Long id, ScheduleRequestDTO dto) {
-        if (dto.getAuthor() != null) {
-            scheduleDao.updateAuthor(id, dto.getAuthor());
+
+        validatePassword(id,dto);
+
+        if (dto.getAuthorId() != null) {
+            authorService.validateAuthor(dto);
+            scheduleDao.updateAuthor(id, dto.getAuthorId());
         }
         if (dto.getTask() != null) {
             scheduleDao.updateTask(id, dto.getTask());
         }
-        return scheduleDao.getScheduleById(id);
+        return scheduleDao.getScheduleById(id).get();
     }
 
     @Override
-    public void deleteSchedule(Long id) {
-        scheduleDao.deleteSchedule(id);
+    public void deleteSchedule(Long id, ScheduleRequestDTO dto) {
+        Optional<ScheduleResponseDTO> optional = scheduleDao.getScheduleById(id);
+        if(optional.isPresent()) {
+            validatePassword(id, dto);
+            scheduleDao.deleteSchedule(id);
+        }
+        else {
+            throw new ScheduleNotFoundException(id + "은 없는 아이디 입니다.");
+        }
+
+    }
+
+    private void validatePassword(Long id, ScheduleRequestDTO dto) {
+        if (!dto.getPassword().equals(scheduleDao.getPasswordById(id))){
+            throw new PasswordMissMatchException("비밀번호가 일치하지 않습니다.");
+        }
     }
 }
